refactor: Adding src folder
