+++ __tests__/unit/rendering/diagram_rendering.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 
 import { dia, shapes } from 'jointjs';
-import { ViewRenderer, ViewSettings } from '../../../index';
+import { ViewRenderer, ViewSettings } from '../../../src';
 import { checkModelsEquality } from '../utils/equality_checker';
 import referenceView from '../data/complete/reference_view.json';
