+++ package.json
@@ -33,7 +33,7 @@
     "test": "mocha --require ts-node/register --require tsconfig-paths/register ./__tests__/**/*.test.ts",
     "test:unique": "mocha --require ts-node/register --require tsconfig-paths/register ./__tests__/**/*.test.ts --grep",
     "test:cov": "nyc mocha --require ts-node/register --require tsconfig-paths/register ./__tests__/**/*.test.ts",
-    "build": "npx babel lib --extensions \".js,.ts\" --out-dir dist --copy-files --no-copy-ignored && tsc --watch"
+    "build": "npx babel src --extensions \\\".js,.ts\\\" --out-dir dist --copy-files --no-copy-ignored && tsc --watch"
   },
   "license": "MIT",
   "keywords": [
