+++ src/lib/common/enums/connectors.ts
