+++ src/lib/common/enums/glyphSymbol.ts
