+++ src/lib/common/enums/nodeShapeClassification.ts
