+++ src/lib/common/enums/nodeType.ts
