+++ src/lib/common/enums/pointerType.ts
