+++ src/lib/glyphGenerator/glyphsDescriptions/glyphs.ts
