+++ src/lib/model/GraphOutput.ts
