+++ src/lib/viewRenderer/nodeRendering/ShapeBuilder.ts
