+++ src/lib/viewRenderer/utils/archimateDomainUtils.ts
