+++ src/lib/viewRenderer/utils/styleUtility.ts
